chore: 파일 이름의 혼동 방지를 위해 파일명 변경
- 승준님께서 제작하신 authStore와의 혼동 방지를 위해 AdminAuthStore로 변경
